@@ -16,7 +16,7 @@ Gem::Specification.new do |s|
   s.files = Dir['{app,config,db,lib}/**/*'] + ['Rakefile', 'README.md']
   s.test_files = Dir['spec/**/*']
 
-  s.add_dependency 'rails', '~> 3.2.0'
+  s.add_dependency 'rails', '~> 4.0.0.rc2'
 
   if RUBY_PLATFORM =~ /java/
     s.add_development_dependency 'jdbc-sqlite3'
