@@ -9,32 +9,32 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20121231010454) do
+ActiveRecord::Schema.define(version: 20121231010454) do
 
-  create_table "identities", :force => true do |t|
-    t.string   "username"
+  create_table "identities", force: true do |t|
+    t.string   "uid"
     t.string   "token"
     t.string   "account_type"
     t.integer  "account_id"
     t.string   "reset_token"
     t.string   "remember_token"
     t.string   "type"
-    t.datetime "created_at",     :null => false
-    t.datetime "updated_at",     :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  add_index "identities", ["remember_token"], :name => "index_identities_on_remember_token"
-  add_index "identities", ["reset_token"], :name => "index_identities_on_reset_token"
-  add_index "identities", ["username"], :name => "index_identities_on_username"
+  add_index "identities", ["remember_token"], name: "index_identities_on_remember_token"
+  add_index "identities", ["reset_token"], name: "index_identities_on_reset_token"
+  add_index "identities", ["uid"], name: "index_identities_on_uid"
 
-  create_table "users", :force => true do |t|
+  create_table "users", force: true do |t|
     t.string   "email"
     t.string   "session_token"
     t.text     "dismissed_helpers"
-    t.datetime "created_at",        :null => false
-    t.datetime "updated_at",        :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
 end
